Clarify data types
- Prefer fixed width integer int8_t over Char (both use 1 byte).
- Int stores a 16-bit (2-byte) value with a range of -32,768 to 32,767. For numbers that are only positive, replace with the smallest fixed width type.

+++ apprentice-wizard/Ready.h
@@ -8,8 +8,8 @@ extern Arduboy2 arduboy;
 class Ready : public Scene
 {
   public:
-  int counter;
-  int stage;
+  uint8_t counter;
+  uint8_t stage;
   Game* game;
   void init(Game *pgame);
   virtual void init(){}; // WARNING: not use default init()
